Rename Energy_consumption.ipynb to Multiple_Linear_Regression_Projects/Energy_consumption.ipynb
The objective of this project is to predict energy consumption (the target variable) based on features such as building type, square footage, number of occupants, number of appliances used, average temperature, and day of the week. Some of these features are categorical; therefore, the OneHotEncoder and LabelEncoder functions from the scikit-learn library were used to properly preprocess the data.

At the end, the model was evaluated using a comparative approach between the predicted values and the actual values from the test set. After testing the robustness of the model, an example case was used to predict new values.